+++ src/x11_monitor.c
@@ -179,8 +179,8 @@ void _glfwPollMonitorsX11(void)
             {
                 if (screens[j].x_org == ci->x &&
                     screens[j].y_org == ci->y &&
-                    screens[j].width == ci->width &&
-                    screens[j].height == ci->height)
+                    screens[j].width == (int)ci->width &&
+                    screens[j].height == (int)ci->height)
                 {
                     monitor->x11.index = j;
                     break;
@@ -572,7 +572,7 @@ void _glfwSetGammaRampX11(_GLFWmonitor* monitor, const GLFWgammaramp* ramp)
 {
     if (_glfw.x11.randr.available && !_glfw.x11.randr.gammaBroken)
     {
-        if (XRRGetCrtcGammaSize(_glfw.x11.display, monitor->x11.crtc) != ramp->size)
+        if (XRRGetCrtcGammaSize(_glfw.x11.display, monitor->x11.crtc) != (int)ramp->size)
         {
             _glfwInputError(GLFW_PLATFORM_ERROR,
                             "X11: Gamma ramp size must match current ramp size");
